Subsitutions sometimes did not go under Forall binder
In capture-avoiding type substitutions, we do a check to see if that which we substitute may cause capture. In the case that it did cause capture, we freshened the binder, and recursively substituted the body.

However, in the case where no capture would happen, we forgot to substitute the body recursively.